+++ package.json
@@ -9,6 +9,7 @@
     "react-dom": "^16.2.0",
     "react-flexbox-grid": "^2.0.0",
     "react-router": "^4.2.0",
+    "react-router-dom": "^4.2.2",
     "react-scripts": "1.0.17",
     "styled-components": "^2.4.0"
   },
