+++ src/App.js
@@ -1,31 +1,32 @@
 import React, { Component } from "react";
+import { render } from "react-dom";
+import { BrowserRouter, Route } from "react-router-dom";
 
 import "normalize.css";
 import "flexboxgrid2";
 import "./App.css";
 
 import Header from "./Header";
-import Explore from "./Explore";
-import Experiences from "./Experiences";
+
+import Landing from "./Landing";
 import Homes from "./Homes";
-import Popular from "./Popular";
-import Featured from "./Featured";
-import Footer from "./Footer";
 
 class App extends Component {
   render() {
     return (
-      <div>
-        <Header />
-        <Explore />
-        <Experiences />
-        <Homes />
-        <Popular />
-        <Featured />
-        <Footer />
-      </div>
+      <BrowserRouter>
+        <div>
+          <Header />
+          <Route path="/" exact component={Landing} />{" "}
+          {/* отрендерить компонент Landing когда мы находимся на / */}
+          <Route path="/homes" component={Homes} />{" "}
+          {/* отрендерить компонент Homes когда мы находимся на /homes/* */}
+        </div>
+      </BrowserRouter>
     );
   }
 }
+
+render(<App />, document.getElementById("root"));
 
 export default App;
